Continuous space: float positions, 24x24 natural map, real contour lines
Positions are now floating-point world coordinates instead of grid cells.
Movement is a straight-line dash to any point within range whose segment
stays clear of rocks and cliffs, and units are solid bodies that cannot
overlap. Attacks stay pointer-based and enumerable.

Rules API reshaped around the continuous move space: IsLegalMoveTarget is
the authority, ProjectMove clamps an arbitrary request onto the legal
region (constraint projection in place of hard masking), GetMoveRegion
casts 72 rays for the UI/sampler, and SampleLegalMoves draws validated
samples. GetAllLegalActions is explicitly no longer exhaustive. Ranges,
LOS, and damage all work off Euclidean geometry via a new Geometry
module (Amanatides-Woo tile traversal).

Map: 24x24 generated offline with fractal noise, symmetrized, quantized
to four elevation bands and repaired for reachability, then baked in;
12 units per side. Relief now renders as true topographic isolines
(marching squares) with spot heights on summits and heavy contours on
cliffs, replacing the tile-tinted concentric rings.

Window is resizable/maximizable with F11 fullscreen; the board reframes
on resize and the HUD moved into the margins. Log schema v4 (float
coordinates) - grid-era logs v1-v3 are not training-compatible.

27 tests green in dotnet and Unity EditMode, including projection
soundness over arbitrary requests; autoplay logs validated against v4.

+++ Assets/Scripts/Core/Unit.cs
@@ -19,11 +19,12 @@ namespace Tactix.Core
         [JsonConverter(typeof(StringEnumConverter), true)] // camelCase string
         public UnitType Type { get; set; }
 
+        /// <summary>World position (continuous). Tile (i,j) is centred at (i,j).</summary>
         [JsonProperty("x")]
-        public int X { get; set; }
+        public double X { get; set; }
 
         [JsonProperty("y")]
-        public int Y { get; set; }
+        public double Y { get; set; }
 
         [JsonProperty("hp")]
         public int Hp { get; set; }
